Rename speculative to gamed across all component and file names
